Changed service protocol: sessions are now created by parent This way, the parent can control how many session a child can have at each service. Additionally, we don't need the connection anymore because the parent can directly create the session.
This is achieved by giving each child its own service-portals in
the parent, so that we can call a portal of the service without
having to trust him. If the service does not respond but stays
in the session-create-portal for ever, only this child is dead.
I.e. only the child that creates the session needs to trust
the service, which is required anyway.
But it's not as easy as it seems: we have to be really careful
to not hold locks during that time. This can get quite ugly :/