Unify tools/novaboot between NUL and NRE
This updates novaboot to the version found in the NUL repository. This
version contains most of the fixes needed for NRE (I'm able to boot NRE
on my test box with pulsar bootloader).

The fixes not included are namely the last two changes from
ebd6994068140c3c4a950d898568dc3e5c346320 because the first one would
break my setup and the second is IMHO not needed.