Changed session-/service-protocol.
I've decided to do the session-creation similar to Genode, i.e.
the parent creates and destroys sessions at the service. After
thinking about and trying various other schemes, this seems to
be the only reasonable one if the parent should be able to
restrict session-creations of its childs. Because otherwise
one can't control session-parameters that easily if at all.
And this is critical IMO for things like giving a child only
access to a subtree of the filesystem. So, restricting only
the number of sessions isn't enough (which can be achieved
in a simpler way). Additionally, the whole protocol is simple.

The disadvantage is that currently the parent has to trust
the services (its childs) to a certain degree because he does
an IPC-call to them. IMO this can be solved by introducing
IPC timeouts and deadlock detection in NOVA.