Rewrote the destruction of sessions and childs.
Now we have a general algorithm for deleting objects that may be
used on multiple CPUs. The basic idea is to invalidate an object
first, ensure that nobody is using it anymore and delete it finally.
For the two usecases that exist so far, the second step is achieved by
letting all affected Ecs call through an empty portal.

Additionally, the session portals and child portals receive directly
the pointer to the session and child object, respectively. This avoids
that we have to find that object in an efficient way from the selector.
The other consequence is thus that the number of sessions/childs is no
longer statically limited but arbitrary large (limited by the amount of
memory).

Furthermore, RCU isn't used for the destruction anymore, but we have
a class Reference now that allows us to get a session/child by id and
use it without the risk of it being deleted during that time. So, in
the portals these objects can be used as they are because they will
never be destroyed during an active portal call, but when one wants to
use them at other locations, one uses Reference (which might destroy
it in the destructor).