Added GlobalThread::join().
This is required to allow to wait for thread termination in a
reasonable way. That is, e.g. before terminating the main-thread
and thus the Pd, we want to wait for all other threads to
terminate to prevent that these threads cause pagefaults or
similar when they weren't completely dead yet.