Replaced the internal vancouver with seoul.
This is done by adding seoul as a git submodule and building it
as a library. This allows us to use it for other purposes
(e.g. VESA bios calls) as well. The NRE specific part has been
changed to match seouls requirements. It is thus at the same
feature-level as before. Although a few things aren't optimal
yet. For example, the console backend simply copies the guest
VGA memory to the host VGA memory in a certain interval. And
the disk backend has to copy to a different DMA descriptor
representation before going to the storage-service.