Don't use the gcc-provided stuff, but our own (fixes #24).
That is, we have our own linker-scripts now, use -nostartfiles
and provide them ourself including calling __register_frame and
the global constructors.
This prevents that we need to write to the cross-compiler-folder,
avoids the eh_frame-problem in crtend.o of gcc we had previously.
Using our own linker-scripts gives us more control and removes
stuff we didn't need anyway.