Use Reference when working with Thread.
Since GlobalThreads destroy themselfs on terminate, it was a bit
unfortunate previously that one had to be careful when accessing
the thread-object after the thread had been started. Now, we
don't pass a pointer to a thread-object around but a Reference
to it. Thus, the thread-object will only be destroyed if the last
Reference for it is destroyed.