Fixed constructor calls for clang.
As it seems, clang generates different constructors depending on
whether you build for x86_32 or x86_64. The reverse ones are
used (and required) on x86_32, but not on x86_64. Worse, on x86_64
clang is unable to compile the for-loop correctly. It doesn't
check the condition before the loop and since the array empty it
calls whatever there is at that address. WTF?