Added a tool for automatic testing.
It runs multiple boot-scripts in qemu and bochs for various
configurations (debug/release, x86_32/x86_64, gcc/clang) in
parallel and checks whether they were successfull.
To do that, I also had to change novaboot to make it use a
temp-file for the menu.lst instead of a file at a fixed
location. And I had to adjust the disk-test to make it
automatically testable.